release(canvas): 0.1.13
Publish Pulse Canvas 0.1.13 artifacts and keep canvas workspace file-size governance green via mechanical helper extractions.
